Graphs

GCMMesh.Graphs.Graph — Type

Graph{I,FT,IA1D} <: AbstractGraph

Connectivity graph for a mesh.

Fields

nverts

number of graph vertices

edge_data

edge data for each vertex

edge_offset

edge data offsets for each vertex

vertices

vertex numbers

spectral_radius

spectral radius estimate/bound of graph laplacian matrix

GCMMesh.Graphs.build_vertex_graph — Function

build_vertex_graph(mesh::AbstractMesh{I,FT}) where {I,FT}

This function builds a graph, with adjacency based on mesh elements connected by a vertex.

GCMMesh.Graphs.build_face_graph — Function

build_face_graph(mesh::AbstractMesh{I,FT}) where {I,FT}

This function builds a graph, with adjacency based on mesh elements connected by a face.

GCMMesh.Graphs.laplacian_x_v! — Function

laplacian_x_v!(vout::FTA1D, vin::FTA1D, grph::Graph)

This function computed the product of graph Laplacian matrix with vector vin.

GCMMesh.Graphs.fiedler_vector — Function

fiedler_vector(graph::Graph{<:Any,FT}, ::Lanczos)

This function estimates the Fiedler vector using Lanczos method.

Reference: Estimating the Largest Eigenvalue by the Power and Lanczos Algorithms with a Random Start Kuczynski, Jacek; Wozniakowski, Henryk

https://academiccommons.columbia.edu/doi/10.7916/D8B56SVV

fiedler_vector(graph::Graph{<:Any,FT}, ::PowerIt)

This function estimates the Fiedler vector using power iteration.

fiedler_vector(graph::Graph{<:Any,FT}, ::Direct)

This function computes the Fiedler vector using Julia/LAPACK subroutines. This is intended for testing purposes only. Lanczos method is the recommended default.

GCMMesh.Graphs.laplacian_matrix_full — Function

laplacian_matrix_full(graph::Graph{<:Any,FT})

Computed the full Graph laplacian matrix explicitly. This is recommended for testing purposes only.

GCMMesh.Graphs.split_graph — Function

split_graph(graph::Graph{I,FT}, n1)

This function partitions the graph such that the first partition has n1 vertices and second partition has graph.nverts - n1 vertices.