Mesh

GCMMesh.Mesh.Mesh1D — Type

Mesh1D{I,IA1D,IA2D,FT,FTA2D,SNT,NB} <: AbstractMesh{I,FT}

Conformal mesh for a 1D manifold. The manifold can be embedded in a higher dimensional space.

For 1D meshes, faces are identical to meshes.

GCMMesh.Mesh.Mesh2D — Type

Mesh2D{I,IA2D,FT,FTA2D} <: AbstractMesh{FT}

Conformal mesh for a 2D manifold. The manifold can be embedded in a higher dimensional space.

                    Quadrilateral

            v4            f3           v3
              o------------------------o
              |                        |      face    vertices
              |       <----------      |             
              |                 |      |        f1 =>  v1 v2 
           f4 |                 |      | f2     f2 =>  v2 v3
              |                 |      |        f3 =>  v3 v4
              |                 |      |        f4 =>  v4 v1
              |       -----------      |
              |                        |
              o------------------------o
             v1           f1           v2

Reference: https://gsjaardema.github.io/seacas-docs/html/element_types.html#ordering

Fields

nverts

number of unique vertices in the mesh

nfaces

number of unique faces in the mesh

nelems

number of elements in the mesh

nbndry

number of zones in the mesh

coordinates

x₁, x₂, ... coordinates of vertices (nverts, dim), dim can be greater than 2 for 2D manifolds embedded in higher dimensional space

unique_verts

unique vertices (n_uniquevertices)

uverts_conn

connectivity information for unique vertices

uverts_offset

offset information for uvertsconn `(nunique_verts + 1)`

face_verts

face vertices numbers (nfaces, 2)

face_neighbors

boundary elems for each face (nfaces, 5) -> [elem1, localface1, elem2, localface2, relative orientation]

face_boundary

face numbers on each boundary (nfaces)

boundary_tags

boundary tags

boundary_tag_names

boundary tag names

face_boundary_offset

face boundary offset for each boundary for face_boundary array

elem_verts

vertices numbers for each elem (nelems, 4)

elem_faces

face numbers for each elem (nelems, 4)

GCMMesh.Mesh.equispaced_line_mesh — Function

equispaced_line_mesh(
x1min::FT,
x1max::FT,
n1,
per = false,

) where {FT<:AbstractFloat}

This function generates an equispaced line mesh with user provided coordinates for left and right domain vertices, number of mesh elements and periodicity information.

GCMMesh.Mesh.equispaced_rectangular_mesh — Function

equispaced_rectangular_mesh(
x1min::FT,
x2min::FT,
x1max::FT,
x2max::FT,
n1,
n2,
per = (false, false),

) where {FT<:AbstractFloat}

This function generates an equispaced rectangular mesh with user provided coordinates of left bottom and right top vertices, number of mesh elements in each direction and periodicity information.

GCMMesh.Mesh.cube_panel_mesh — Function

cube_panel_mesh(ne)

This function builds a cube panel mesh with a resolution of ne elements along each edge.

           v8 (xs,xe,xe)          v7 (xe,xe,xe)
             o--------e11---------o       
            /|                   /|                                   
           / |                  / |       
          /  |                 /  |       
        e12  e8               e10 e7       
        /    |               /    |       
       /     |            v6/     |       
   v5 o--------e9----------o      |       
      |    v4o------e3-----|------o v3 (xe,xe,xs)
      |     /   o------->  |     / 
      |    /   /           |    / 
      e5  e4  /           e6   e2 
      |  /   /             |  / 
      | /   /              | / 
      |/   o               |/ 
      o--------e1----------o   
     v1                    v2 
   (xs,xs,xs)               (xe,xs,xs)

   panel 1 => 1 4 3 2
   panel 2 => 2 3 7 6
   panel 3 => 3 4 8 7
   panel 4 => 1 5 8 4
   panel 5 => 1 2 6 5
   panel 6 => 5 6 7 8

   edge  1 => 1 2 
   edge  2 => 2 3
   edge  3 => 3 4
   edge  4 => 4 1
   edge  5 => 1 5
   edge  6 => 2 6
   edge  7 => 3 7
   edge  8 => 4 8
   edge  9 => 5 6
   edge 10 => 6 7
   edge 11 => 7 8
   edge 12 => 8 5

        v8 +---e11---+ v7
           | ^       |
           | |       |
          e8 ^   3   e7 
           | |       |
        v4 | o--<-o  | v3      v7        v8

v8 +–-e8––+–-e3––+––e7–-+–-e11–-+ | ^ | o–>-> | o–>–> | ^ | | | | | | | | | | e12 ^ 4 e4 ^ 1 e2 ^ 2 e10^ 6 e12 | | | | | | | | | | o-<-<-o | o | o | o<-<–o | +–-e5––+–-e1––+––e6–-+––e9–-+ v5 v1| o–>–o | v2 v6 v5 | | | e5 5 | e6 | ↓ | | | +–-e9––+ v5 v6

GCMMesh.Mesh.sphere_mesh — Function

sphere_mesh(ne, radius::FT, wtype::AbstractSphereWarp) where {FT<:AbstractFloat}

Generates a sphere mesh of a given radius, warp type and number of elements along each side of a cubed surface. The sphere mesh is generated by applying the warping function to a cube panel mesh.

GCMMesh.Mesh.cubed_sphere_warp — Function

cubed_sphere_warp(::EquiangularSphereWarp, a, b, c)

Given points (a, b, c) on the surface of a cube, warp the points out to a spherical shell of radius R based on the equiangular gnomonic grid proposed by [Ronchi1996]

Returns a tuple of (x1, x2, x3), where x1, x2, x3 are the corresponding coordinates on a sphere of radius R.

The "Cubed Sphere": A New Method for the Solution of Partial Differential Equations in Spherical Geometry C. RONCHI,,1 R. IACONO, AND P. S. PAOLUCCI† JOURNAL OF COMPUTATIONAL PHYSICS 124, 93–114 (1996) https://github.com/CliMA/ClimateMachine.jl/blob/master/src/Numerics/Mesh/Topologies.jl

cubed_sphere_warp(::EquidistantSphereWarp, a, b, c)

Returns a tuple of (x1, x2, x3), where x1, x2, x3 are the corresponding coordinates on a sphere of radius R.

https://github.com/CliMA/ClimateMachine.jl/blob/master/src/Numerics/Mesh/Topologies.jl